[events] - Filter Tags
Update the filterTags function to return either UA pageview, UA event,
GA4 config, or GA4 event tags.

Adding the listTagNamesAndIds function to extract tag names and IDs and
put them in a double array. This array can then be easily written to a
the spreadsheet.

Changed the entityId variable to analytivsVersion for clarity.

Created the uaTagType variable.

Change-Id: I1263aafdde560c433090178f9606e0999f5e2de8