[events] - Update getTagsFromSheet
Updated the getTagsFromSheet function to call a list of all tags from
the GTM container and then filter those tags to create an array that
only contains the tags that were selected to be migrated in the sheet.
This should be more efficient from an API perspective (and avoid QPS
limits) since it only requires one API request rather than multiple. The
function also now returns the selected config tag name for an event tag
if UA event tags are being migrated.

Change-Id: I35ffbcf2248c1d5219abf4e025a38e0480ff95ac