[filtering] - Modification of Tag Filtering
This is primarily a change of filterTags() to provide a more advanced
set of tag filtering options. This update mainly changes how tags are
filtered and then written on the sheet:
- When listing tags, only those UA event and pageview tags that use the
selected GA settings variable will be displayed
- When listing custom definitions, only those custom definitions that
exist within the UA event or pageview tags and settings
variable that are being migrated will be displayed
- When listing fields to set, only those fields that exist in the UA
pageview tags and settings variable that are being migrated will be
displayed
- When listing event data, only the category, action, and label data for
those events that are being migrated will be displayed

This also adds the settings variable to the event migration flow.

Code now made redundant for event data and listing custom dimensions
was removed.

Change-Id: Ia7d9407b4f16c9228745fe8570efa7347f547396