[error handling] -  Append GA4 to names when duplicate
Added checkForDuplicateTagName(), which checks if the supplied tag name
already exists in the workspace. If it does, then ' - GA4' is appended
to the end of the tag name. If not, then the supplied tag name is used.

Increased the write request delay to 4000 milliseconds to be in
compliance with the default QPS for the API.

Change-Id: Ie002bfdacc7113391a617b5ab142ae3162bc721c