[measurement] - Update Tool Measurement
Added measurementConsentCheck() to onOpen(). The tool will now check if
measurement consent has been requested when the sheet first opens.

Also fixed a small bug with writeFieldsToSheet() where it checks if
there are more than 0 fields to write to the sheet before attempting
to writhe the fields to the sheet.

Change-Id: Ia570fea215567f9f430c2e64f1cd1e4196c0c167