[caching] - Add Caching for list requests
Added caching of the variable and tag list API requests to decrease the
number of requests made to the server. The data is cached for five
minutes. This can be adjusted by changing the value of the cacheTimeout
variable.

Several references to getKey(), getMap(), etc. were changed to key, map,
etc. to work with the caching.

Change-Id: I652625a7a6c43679e9dab79d41cf33115f685a79