[update] - Workspace and GTM Variable Listing
This update does three main things refactor the code, add workspace
listing, and add GTM variable listing.Instead of storing all of the code
in a single file, this update breaks various functions into particular
files. Some of the functions have also been re-written and simplified.

Previously, the script relied on a user entering their workspace URL to
run correctly. This introduced the potential for error. The script
now lists GTM workspaces and a user can select from the list.

The script can now list GTM variables, both user and built-in, and
sets them as a dropdown list in various areas in the spreadsheet.

Change-Id: I8a248aa22a462d1981bf1658365ffd591ee8bf69