[pageview] - Fix Pageview Event Creation
This update fixes a bug that was preventing pageview event tags from
being created. It also enables the user to specify which config tag
should be used when creating a pageview event tag.

The checkRelease() function was also updated to check for new releases
of the script. If a new release is found, then the user will be prompted
to make a new copy of the template spreadsheet. This check is run every
time a user writes UA settings variables to the spreadsheet. If a user
dismisses the prompt, then it will not be shown again.

Change-Id: I51f55fd66e62420f7fe7a4f7607369f4601719fe